+++ dali/kernels/audio/mel_scale/mel_filter_bank_gpu_test.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,8 +14,10 @@
 
 #include <gtest/gtest.h>
 #include <random>
-#include <dali/test/tensor_test_utils.h>
-#include <dali/test/test_tensors.h>
+#include <tuple>
+#include <vector>
+#include "dali/test/tensor_test_utils.h"
+#include "dali/test/test_tensors.h"
 #include "dali/kernels/audio/mel_scale/mel_filter_bank_gpu.h"
 #include "dali/kernels/audio/mel_scale/mel_filter_bank_test.h"
 #include "dali/kernels/kernel_manager.h"
@@ -175,7 +177,7 @@ INSTANTIATE_TEST_SUITE_P(MelScaleGpuTestpuTest, MelScaleGpuTest, testing::Combin
     testing::Values(5000.0f, 8000.0f),  // fmax
     testing::Values(0, 2, 3)));  // axis
 
-}
-}
-}
-}
+}  // namespace test
+}  // namespace audio
+}  // namespace kernels
+}  // namespace dali
