+++ dali/kernels/audio/mel_scale/mel_filter_bank_test.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -75,7 +75,7 @@ std::vector<std::vector<float>> ReferenceFilterBanks(int nfilter, int nfft, floa
   return fbanks;
 }
 
-}
-}
-}
-}
+}  // namespace test
+}  // namespace audio
+}  // namespace kernels
+}  // namespace dali
